Cast sizeof to long
Reviewed-on: http://review.couchbase.org/24274
Tested-by: Trond Norbye <trond.norbye@gmail.com>

+++ debug/sizes.c
@@ -20,8 +20,10 @@
 int main(void) {
 
     fprintf(stderr, "Sizes\n");
-    fprintf(stderr, "  lcb_t        : %ld\n", sizeof(struct lcb_st));
-    fprintf(stderr, "  lcb_server_t : %ld\n", sizeof(struct lcb_server_st));
+    fprintf(stderr, "  lcb_t        : %ld\n",
+            (long)sizeof(struct lcb_st));
+    fprintf(stderr, "  lcb_server_t : %ld\n",
+            (long)sizeof(struct lcb_server_st));
     return 0;
 
 }
